requirements.txt
alembic==1.15.2
async-timeout==5.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mysql-connector-python==9.3.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
redis==6.1.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
